refactor: adjust code location
