+++ vm/src/stdlib/thread.rs
@@ -307,7 +307,7 @@ pub(crate) mod _thread {
     #[cfg(not(target_arch = "wasm32"))]
     #[pyfunction]
     fn interrupt_main(signum: OptionalArg<i32>, vm: &VirtualMachine) -> PyResult<()> {
-        crate::stdlib::signal::_signal::set_interrupt_ex(signum.unwrap_or(libc::SIGINT), vm)
+        crate::signal::set_interrupt_ex(signum.unwrap_or(libc::SIGINT), vm)
     }
 
     #[pyfunction]
